Added Value Request for Title
Previously, if the title couldn't be fetched, the bookmark would get titled with "Unknown Title" or the host of the URL, but now, the intent will ask the user to type out a title of their choice.
The same goes for when auto-title is disabled and no title is entered.