Fixed Scroll Performance Issue
I had accidentally left in the code to fetch metadata even after the cached version of it was found. This was causing the LazyVStack to reload the view every time I dragged the ScrollView, causing significant performance drop offs.

Now, it is only fetches the images if there is no cached version, restoring the silky smooth scroll of previous Linkeeper versions.