Favorites now correctly responds to drop interactions
When the BookmarksView was previously showing favorites, when there was at least one favorite bookmark, and a new URL was dropped into favorites, it would get added to All Bookmarks but would not be marked favorite.

This is no longer the case; dropping it into the favorites section will infact mark it as favorite, as well as add it to the All section.