Improved (Re)Ordering of Folders
The order of folders in the "All Bookmarks" section now follows the logical order of pinned folders followed by the rest of the folders, that then following the order of the folders before any of them were pinned.
A folder, when unpinned, also now goes directly to the bottom of the list.